refactor: Own the workflow test case schema in the loader
WorkflowTestCaseValidator is imported only by evals/workflows/, but sat in
shared/types.ts, which also pulled zod into a module both eval suites import.
That split the workflow schema across three files, with the workflow-specific
piece furthest from workflows/.

Moves the schema next to the array/uniqueness schema in test_cases_loader.ts
and drops the "backwards compatibility" re-export, which had no consumer.

loadTestCases() now takes the absolute path from resolveTestCasesPath() rather
than re-resolving it, and drops the existsSync guard: readFileSync already
reports a missing file and names the path.
