+++ evals/workflows/test_cases_loader.ts
@@ -1,23 +1,50 @@
 /**
- * Test case loader for workflow evaluations.
- * Uses the shared JSON reader plus workflow-specific validation.
+ * Test case loading and validation for workflow evaluations.
+ * Uses the shared JSON reader plus the workflow test case schema.
  */
 
-import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 import { z } from 'zod';
 
 import { loadTestCases as loadTestCasesShared } from '../shared/test_case_loader.js';
-import { WorkflowTestCaseValidator } from '../shared/types.js';
-import type { WorkflowTestCase } from '../shared/types.js';
-
-// Re-export WorkflowTestCase type for backwards compatibility
-export type { WorkflowTestCase } from '../shared/types.js';
 
 /** The canonical test cases file, resolved from this module so cwd cannot change it. */
 export const DEFAULT_TEST_CASES_PATH = path.join(path.dirname(fileURLToPath(import.meta.url)), 'test_cases.json');
+
+/**
+ * A workflow test case (multi-turn agent conversation).
+ *
+ * Validated rather than cast because it is read from a JSON file and, once synced,
+ * from a Langfuse dataset item that can be edited outside this repo.
+ *
+ * Strict: an unknown key is a typo'd harness knob. Stripping it silently would let
+ * `failTool` sync as nothing, the forced INTERNAL_ERROR never fire, and the eval go
+ * green without testing what it claims to test.
+ */
+export const WorkflowTestCaseValidator = z.strictObject({
+    /** Unique test case ID */
+    id: z.string().min(1),
+    /** Category for grouping (e.g., "search-actors", "call-actor") */
+    category: z.string().min(1),
+    /** User query/prompt given to the agent */
+    query: z.string().min(1),
+    /** Requirements the judge scores the conversation against */
+    reference: z.string().min(1),
+    /** Maximum number of turns allowed (defaults to the config value) */
+    maxTurns: z.number().int().positive().optional(),
+    /** Tools to enable for this test (e.g., ["actors", "docs", "apify/rag-web-browser"]) */
+    tools: z.array(z.string()).optional(),
+    /**
+     * Tool names the harness force-fails with a synthetic INTERNAL_ERROR carrying the real
+     * report-problem nudge. Lets an eval deterministically throw a nudge-eligible error
+     * that the live server + API cannot reproduce on demand. See mcp_client.ts.
+     */
+    failTools: z.array(z.string()).optional(),
+});
+
+export type WorkflowTestCase = z.infer<typeof WorkflowTestCaseValidator>;
 
 const WorkflowTestCasesValidator = z
     .array(WorkflowTestCaseValidator)
@@ -29,24 +56,20 @@ const WorkflowTestCasesValidator = z
  * Absolute path of the file a run reads.
  *
  * `--test-cases-path` is resolved against cwd, like any other CLI path argument.
- * Resolving here rather than at each use site is what keeps the existence check,
- * the read and the dataset name pointing at one file: the shared reader resolves
- * relative paths against `evals/`, so an unresolved path would be checked in one
- * place and read from another.
+ * Resolving here rather than at each use site is what keeps the read and the dataset
+ * name pointing at one file: the shared reader resolves relative paths against
+ * `evals/`, so an unresolved path would name one file and read another.
  */
 export function resolveTestCasesPath(filePath?: string): string {
     return filePath ? path.resolve(filePath) : DEFAULT_TEST_CASES_PATH;
 }
 
 /**
- * Load workflow test cases from a JSON file, validating every case.
+ * Load workflow test cases, validating every case.
+ *
+ * Takes an absolute path from `resolveTestCasesPath`. A missing file surfaces as the
+ * reader's own ENOENT, which already names the path.
  */
-export function loadTestCases(filePath?: string): WorkflowTestCase[] {
-    const testCasesPath = resolveTestCasesPath(filePath);
-
-    if (!fs.existsSync(testCasesPath)) {
-        throw new Error(`Test cases file not found: ${testCasesPath}`);
-    }
-
+export function loadTestCases(testCasesPath: string): WorkflowTestCase[] {
     return WorkflowTestCasesValidator.parse(loadTestCasesShared(testCasesPath).testCases);
 }
